Financial statements now scrape to database.

+++ finance/db_repository/versions/201310282017_dc13c9437a8_add_company_financia.py
@@ -26,58 +26,58 @@ def upgrade():
 
     # Data from income statement:
     # Sales, aka. total revenue
-    sa.Column("sales", sa.Numeric),
+    sa.Column("sales", sa.Float),
     # Cost of goods sold, aka. cost of revenue, total cost of revenue
-    sa.Column("cost_of_goods_sold", sa.Numeric),
+    sa.Column("cost_of_goods_sold", sa.Float),
     # Net income 
-    sa.Column("net_income", sa.Numeric),
+    sa.Column("net_income", sa.Float),
     # Sharehold net income, aka. net income applicable to common shareholders,
     # income available to common incl. extra items
-    sa.Column("shareholder_net_income", sa.Numeric),
+    sa.Column("shareholder_net_income", sa.Float),
     # Multiplier string from income statement
-    sa.Column("income_statement_multiplier", sa.Numeric),
+    sa.Column("income_statement_multiplier", sa.Float),
 
     # Data from balance sheet:
     # Cash and equivalents, aka. cash and cash equivalents
-    sa.Column("cash_and_equivalents", sa.Numeric),
+    sa.Column("cash_and_equivalents", sa.Float),
     # Current assets, aka. total current assets
-    sa.Column("current_assets", sa.Numeric),
+    sa.Column("current_assets", sa.Float),
     # Short term debt, aka. short term debt/current portion of long term debt,
     # notes payable/short term debt
-    sa.Column("short_term_debt", sa.Numeric),
+    sa.Column("short_term_debt", sa.Float),
     # Accounts payable
-    sa.Column("accounts_payable", sa.Numeric),
+    sa.Column("accounts_payable", sa.Float),
     # Other current liabilities
-    sa.Column("other_current_liabilities", sa.Numeric),
+    sa.Column("other_current_liabilities", sa.Float),
     # Current liabilities, aka. total current liabilities
-    sa.Column("current_liabilities", sa.Numeric),
+    sa.Column("current_liabilities", sa.Float),
     # Long term debut
-    sa.Column("long_term_debt", sa.Numeric),
+    sa.Column("long_term_debt", sa.Float),
     # Multiplier string from balance sheet
-    sa.Column("balance_sheet_multiplier", sa.Numeric),
+    sa.Column("balance_sheet_multiplier", sa.Float),
 
     # Data from cash flow statement:
     # Operating cash flow, aka. net operating cash flow
-    sa.Column("operating_cash_flow", sa.Numeric),
+    sa.Column("operating_cash_flow", sa.Float),
     # Capital expenditures
-    sa.Column("capital_expenditures", sa.Numeric),
+    sa.Column("capital_expenditures", sa.Float),
     # Multiplier string from cash flow statement
-    sa.Column("cash_flow_statement_multiplier", sa.Numeric),
+    sa.Column("cash_flow_statement_multiplier", sa.Float),
 
     # Data from revenue / EPS summary:
     # Earnings per share
-    sa.Column("eps", sa.Numeric),
+    sa.Column("eps", sa.Float),
     
     # Cached computation:
     # Approximate diluted shares outstanding, computed as shareholder_net_income/eps
-    sa.Column("shares_outstanding", sa.Numeric),
+    sa.Column("shares_outstanding", sa.Float),
     # Ranker stats:
-    sa.Column("gross_margin", sa.Numeric),
-    sa.Column("net_margin", sa.Numeric),
-    sa.Column("cash_to_debt_ratio", sa.Numeric),
-    sa.Column("net_cash", sa.Numeric),
-    sa.Column("foolish_flow_ratio", sa.Numeric),
-    sa.Column("cash_king_margin", sa.Numeric),
+    sa.Column("gross_margin", sa.Float),
+    sa.Column("net_margin", sa.Float),
+    sa.Column("cash_to_debt_ratio", sa.Float),
+    sa.Column("net_cash", sa.Float),
+    sa.Column("foolish_flow_ratio", sa.Float),
+    sa.Column("cash_king_margin", sa.Float),
   )
